Find linknames across dependencies.
Use lower level as packages to sequentially walk the import tree.
In each package, loop over files and find links as before.

An alternative approach would be to use the analyzer with a "Fact".
This triggers analysis on the complete tree, but seems hacky.
But when accumulating the result, it could prove to be more correct.